refactor: move wrapLogGroup from core/lib/log to core/lib/actions
